requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-phone-field==1.8.1
django-rest-auth==0.9.5
djoser==2.1.0
coverage==5.3
pycryptodome==3.9.9
qrcode==6.1
Pillow==8.0.1
drf-yasg==1.20.0
pandas==1.1.5
pylint==2.6.0
xlrd==2.0.1
gunicorn==20.0.4
django-heroku==0.3.1
